Add files into keeka package
