+++ src/main/java/keeka/InvalidMarkingException.java
@@ -1,3 +1,5 @@
+package keeka;
+
 public class InvalidMarkingException extends RuntimeException {
     public InvalidMarkingException(String message) {
         super(message);
