+++ src/main/java/keeka/InvalidTaskException.java
@@ -1,3 +1,5 @@
+package keeka;
+
 public class InvalidTaskException extends RuntimeException {
     public InvalidTaskException(String message) {
         super(message);
