+++ src/main/java/keeka/ParsedSaveContent.java
@@ -1,3 +1,5 @@
+package keeka;
+
 public class ParsedSaveContent {
     private char taskCode;
     private boolean isDone;
