+++ src/main/java/keeka/Parser.java
@@ -1,3 +1,5 @@
+package keeka;
+
 import java.io.IOException;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
