+++ src/main/java/keeka/Storage.java
@@ -1,3 +1,5 @@
+package keeka;
+
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileWriter;
