+++ src/main/java/keeka/Task.java
@@ -1,3 +1,5 @@
+package keeka;
+
 public class Task {
     private String description;
     private boolean isDone;
