+++ src/main/java/keeka/Ui.java
@@ -1,3 +1,5 @@
+package keeka;
+
 import java.util.ArrayList;
 
 public class Ui {
